inlude new sklearn parent class for simpler implementation of new models
